+++ flight/targets/FlyingF4/System/pios_board.c
@@ -281,8 +281,8 @@ void PIOS_Board_Init(void) {
 		PIOS_DEBUG_Assert(0);
 	}
 
-	/* Connect flash to the approrpiate interface and configure it */
-	uint32_t flash_id;
+	/* Connect flash to the appropriate interface and configure it */
+	uintptr_t flash_id;
 	if (PIOS_Flash_Jedec_Init(&flash_id, pios_spi_flash_id, 0, &flash_m25p_cfg) != 0)
 		panic();
 	uintptr_t fs_id;
