+++ flight/targets/Freedom/System/pios_board.c
@@ -313,7 +313,7 @@ void PIOS_Board_Init(void) {
 
 #if defined(PIOS_INCLUDE_FLASH)
 	/* Connect flash to the approrpiate interface and configure it */
-	uint32_t flash_id;
+	uintptr_t flash_id;
 	if (PIOS_Flash_Jedec_Init(&flash_id, pios_spi_telem_flash_id, 1, &flash_m25p_cfg) != 0)
 		panic();
 	uintptr_t fs_id;
