+++ flight/targets/Quanton/System/pios_board.c
@@ -272,7 +272,7 @@ void PIOS_Board_Init(void) {
 		PIOS_DEBUG_Assert(0);
 	}
 
-	uint32_t flash_id;
+	uintptr_t flash_id;
 	if (PIOS_Flash_Jedec_Init(&flash_id, pios_spi_flash_id, 0, &flash_mx25_cfg) != 0)
 		panic();
 	uintptr_t fs_id;
